Fix dry_run showing "unknown issue" regardless of earlier match
Previously with the construct "then :" we set a true value and aborted
the if/elif/else evaluations but still returned with "false" causing
other evaluations to run ending in the output of log excerpts regardless
of earlier matching.

Following https://stackoverflow.com/a/43840545 I change the functions to
use bare returns without number arguments returning the exit code of the
last call before return.