Ignore incomplete jobs with reason 'quit'
In this case the worker has been restarted within a reasonable
time (systemd had not had to kill it). This usually happens during
deployment or when the worker is restarted/stopped for other
reasons. I suppose we can ignore these incompletes.

Note that more than 50 % of the incomplete jobs on o3 within the
last 24 hours have the reason 'quit'.