openqa-label-known-issues: Use combined reason+autoinst-log to detect more issues
There can be some search terms which are found only in reason and not in
autoinst-log.txt, sometimes in worker-log.txt which we do not parse so
far. Concatenating the reason into the downloaded log content allows us
to also look for search terms matching the reason string. This for
example is now used for https://progress.opensuse.org/issues/64776 with
the reason `setup failure: Cache service status error: Premature
connection close` which is not in autoinst-log.txt .

Related progress issue: https://progress.opensuse.org/issues/69178