Add a wrapper around jq for better error handling
Issue: https://progress.opensuse.org/issues/95830

Currently we only have this in our logs:

    parse error: Invalid numeric literal at line 1, column 10

This is an error message from jq.
We don't know in which script or line this happened, and we don't
know what the input was.

This commit adds a wrapper around jq that outputs script and line number,
and the input in case jq fails.