openqa-label-known-issues: Remove obsolete handling of poo#61922
We encountered many jobs that have poo#61922 as label from
openqa-label-known-issues. This comes from openqa-label-known-issues that
directly detects openQA jobs without any logs and marks as such. In the
meantime we have the "reason" field which in the case of the mentioned jobs
says something along the lines of "abandoned: associated worker $host:instance
re-connected but abandoned the job" and likely the job was automatically
retriggered at that time by openQA itself so we should not need to handle the
issue anymore in openqa-label-known-issues explicitly.

Related progress issue: https://progress.opensuse.org/issues/96019