Fix SQL queries using "NOW()" with proper timezone
As the openQA database uses UTC by default in timestamps but does not
necessarily specify that we need to specify a proper timezone when using
"NOW()" for correct calculations. This also means that commonly more
jobs are returned. Previously "24 hours" would mostly actually be 23 or
during European summer only 22 hours as OSD+O3 run in CET/CEST.