Improve error handling
Add a trap call to make sure we always get a message on stderr in which script
and which line the error happened. We sometimes get a non-zero exit code but
don't know where the error happened.

I didn't add the trap to openqa-investigate, because practically the whole
code is wrapped in a conditional (investigate || error_count++), and
-e (and trap) is disabled by that.

Also turned several echos into echoerr as it was supposed to land in our
log, not in stdout.

Issue: https://progress.opensuse.org/issues/99519