Prevent execution errors to be swallowed from openqa-label-known-issues
Originally I opted to use stdout for diagnostic output in
openqa-label-known-issues and stderr for the relevant left unreviewed
test failures. This causes potential errors to be hidden in particular
when forwarding the output in the hook scripts.

This commit swaps the use of stdout and stderr in
openqa-label-known-issues so that stderr with diagnostic output can
output info as well as error output which can then be read by callers to
investigate problems while the stdout is passed on in the pipe within
the hook scripts.

Related progress issue: https://progress.opensuse.org/issues/91605