Ensure any bugref in force_result labels to be parsed as a real bugref
If openQA would detect a label in a line within a comment then only a
label would be referenced however for auto-review using force_result
labels I think we should highlight the actual ticket used as reference.
This commit is putting the comment (duplicated) in a separate line so
that openQA will render the comment with an according bugref link, not
the label per se.

Related progress issue: https://progress.opensuse.org/issues/109857