Postpone investigation jobs only if pending jobs are in same cluster
Before this change, investigation jobs are postponed if *any* dependency is
still pending. For instance, a pending chained sibling would cause the
investigation of a job to be postponed. That's not needed so this change
will postpone the investigation only if there are pending jobs in the same
parallel cluster.

See https://progress.opensuse.org/issues/95783#note-64