Convert Bash tests to use bpan-org/test-tap-bash
The ingydotnet/test-more-bash and bash+ libraries are deprecated in
favor of the https://github.com/bpan-org versions.

None of the test logic needed changing at all.
The test bootstrap code in each test was replaced with a single line
that does the same thing with the newer libraries.

The bpan-org/test-tap-bash library does everything test-more-bash did
and more.

The test-tap-bash library is packaged in the `bpan` repo itself, so the
Makefile now clones the bpan-org/bpan repo locally and the tests use
these bundled libraries.