Use test-tap-bash's `try` function
The test-tap-bash library has a `try` function that captures stdout and
stderr and puts them in `$got` and puts the status code in `$rc`.

This pattern is used a lot in the bash tests and using `try` makes the
code quite a bit more readable.

NOTE: the var name `got` (instead of `out`) is used because got/want are
the terms used pervasively throughout test-tap (in error messages and
such) for result/expected.