Cleaner usage of temp files as subprocess flags
The bashplus package provides the `+fs:mktemp` function which creates a
self-deleting temp file and puts the name in `$temp`.

We change the trigger code to put text in the file, and change the test
asssertion to check that the file is not empty.

Overall I think this is a slightly cleaner approach.

It would be nicer to not write to disk at all.
I think this is possible but requires a feature to be added to
test-tap-bash.