Make openqa-label-known-issues take a single url
and provide a wrapper script openqa-label-known-issues-multi for
batch operations.

This makes openqa-label-known-issues easier to use and test.

We can call a single command for a single url that we pass as a
parameter.
Having to pipe the url makes it harder to use.
Also with a command for a single url we can easier write tests for
this command.

Related issue: https://progress.opensuse.org/issues/98862