Make regex more performant
Having several `.*` in a regex with the `s` flag can involve a lot of
backtracking on a logfile, but in reality only one of the .* must be
line spanning.

Looking into the gru logs, this regex was timing out very often.
Removing the s flag and replacing the line spanning .* with
[\S\s]* will make it faster