WIP - Bash code review - first commit
aka - Wrap some long lines in README.md

This is just a benign first commit so that I can create a pull request
for this repository.

Per https://github.com/os-autoinst/scripts/issues/208 I wish to code
review the Bash content and make suggestions on possible improvements.

I haven't done this before so I'm not sure of the best way to go about
it.
I'm thinking to use this pull request, where I will make a commit for
each suggestion with a couple of code changes to demonstrate the
objective.
Then we can can discuss it.

I don't expect the PR to be merged, but you might end up cherry-picking
some commits in it.

Let me know if this approach works for you and I'll start adding commits
soon after.