Add a run-with Makefile function
The run-with function checks to make sure that the runner command is
installed before running the test.

If not installed it issues a warning message.
It used to just fail, which causes all of `make test` to fail.

Also before there where lines to check if shellcheck and yamllint were
installed. They printed a msg but did not stop the commands from
attempting to run only to obviously fail for not being installed.

There was no previous check for py.test which was always failing for me
since I didn't have it installed.

Now when something isn't installed that test is just skipped with an
explanation message.