Remove usage of `env` from Makefile rules
One of the benefits here of setting `SHELL := bash` is that we no longer
need to use env to set environment variables before a shell command.
Bash just handles that itself.

Removed a useless use of `cat`.

Dropped the `-e` from `bash -ex` as that is handled in
./openqa-label-known-issues-multi itsef.

I'm not completely sure about that `grep -v` but it seem like it was
trying to avoid the error from invalid JSON (noted in the comment).
Putting `-` prefix on a rule command tells `make` not to stop if the
command fails. Not sure that is correct here or not.