Various ways to split long lines in Bash
Personally I try to keep all of my coding in any language or format
under 80 columns, but that's just a personal preference.

However, it has allowed me to learn a lot of tricks for splitting lines
in Bash.

In this commit I show a bunch of them. And hopefully you agree then
changes make things more readable.

The best thing to know is that you can split a line after a `|`, `&&` or
`||` token. Bash knows you aren't finished with that command and so no
backslash `\` is required to continue.

Next, `$(...)` forms a code block within. You can put any Bash inside
and it can be multiline. The inner Bash follows all the same rules as
the outer Bash.

The long jq coomands are long strings and bash supports strings to be
multiline. Since jq doesn't mind ws to have newlines, it works the same.
Note I put `|` chars in jq in front rather than at end of lines, because
these `|` have no special meaning to Bash here.

I defined header in 2 commands instead of one, as an example for
splitting a long line with a cleaner look.

`$'...'` is not just a way to quote a newline in Bash. It's a full
quoting syntax like `"..."` or `'...'`. It works like `'...'` but
supports escapes (but not interpolation).
Try `echo $'foo\nbar'`.