Better use of `[[ ... ]]`
I think it's great that you use `[[` instead of `[`.

The former is keyword syntax while that latter acts as a command.
Being syntax allows Bash to be able to treat what's inside specially.

The best thing is you almost never need to use quotes inside `[[`.

You do need to double quote a variable on the RHS for the `==` and `!=`
operators but shellcheck wil catch that.

Speaking of spellcheck is great about catching things that are a problem
but not good at pointing out unrequired syntax. I don't like quoting
things that don't need quotes as it makes people wonder why you quoted
them in the first place. Quoting implies a special semantic intent.

You should use the == != <= etc for string comparison and eq ne le etc
for numbers. Opposite of Perl. Bourne shell came first so good on Perl
for switching it!

Note: I left in `== 200` because I think of http status codes as
numberic strings. I'm honestly not sure how to treat them. Still
learning :)

The `-n` operator is never needed. `[[ $foo ]]` means same as `[[ -n
$foo ]]` and `[[ ! $foo ]]` means same as `[[ -z $foo ]]`.

Changed `[[ $dry_run == 1 ]]` to `[[ $dry_run ]]` as I don't think `1`
is special here.