Use the `x() (...)` function form
If a function has no desirable side-effects (changing directory, setting
global vars) you can define them like:

    fn() ( ... )

instead of

    fn() ( ...; )

and that means the same as

    fn() { ( ... ); }

The nice thing about this is you don't ever need to use `local` because
the subprocess doesn't allow you to change global state.

Note that a one line function declared this way also needs no final
semicolon.

I changed all your functions that didn't have side-effects to use this
form.

I try to use this form unless I need to set state. It ends up leading to
better overall programming decisions.

Note that functions called in a tight loop that need to be fast as
possible should not use this form as subprocesses incur a slight
overhead. Generally this should not be a concern.