Change run-with function to error instead of warn
The 'make test' command will now error when dependencies are not
installed.

For py.test I made it use a virtualenv install of py.test and requests
modules. The Makefile looks for a Python 3 binary and errors otherwise.

I had python3 installed locally but not py.test. I didn't feel like that
should stop me from installing it temporarily on the fly like we do for
bpan.