Fix checking for unknown issues
Issue: https://progress.opensuse.org/issues/128405

The output for unown issues was changed to "Unknown test issue", but the
test and the READE have never been adapted, mainly because we mock
the function in our test, and the mock still uses the old string.

I will think about a way to mock less.