Prevent useless retries on osc cat always failing in pipe
https://github.com/os-autoinst/scripts/pull/305 introduced a problem
causing "osc cat" to always retry as it fails within a piped call.
Apparently there is a SIGPIPE due to the head in the command:

```
osc cat … | grep '…' | head -n1
```

which can be seen with `echo "${PIPESTATUS[@]}"`. The
`grep … | head -n1`
can be simplified to `grep -m1` which gets rid of the sigpipe of grep
but keeps the failure of osc cat. But we can use process substitution to
rescue :)

This also allows us to enforce the proper handling of pipefails
everywhere else.

Related progress issue: https://progress.opensuse.org/issues/157018