Handle errors from the label_on_issues_from_issue_tracker
When `label_on_issues_from_issue_tracker` fails means that it doesn't add a
label on the ticket. Not added a comment it should not signal an ERR which
triggers the trap to be executed.
The current commit tries to solve again the problem which
https://github.com/os-autoinst/scripts/pull/335 solved but introduced a
new bug with a bad design.
After this commit the `label_on_issues_from_issue_tracker` is removed from the
_if_ block which caused early exit from trap when exit code was non-zero.
Redesigned in a simplier readable format keeping the logic the same and included
the requirements intacted. `label_on_issues_from_issue_tracker` is accessed
when autoinst-log exits or reason is found.
In case of some curl issues, no labeling occurs when the job is
old based on a hard-coded variable or when none of the requisites found.
Finally if none of them matches the issues the script will continue
with the next checks in the sequence of actions, as it was before.

https://progress.opensuse.org/issues/165716

Author: perlpunk Tina Müller (tinita)
Signed-off-by: ybonatakis <ybonatakis@suse.com>