Prevent override the autoinst-log context within handle_unreachable
`handle_unreachable` can override the log file passed to the function. This
should not allowed because when the `handle_unreviewed` function is reach will
run with incorrect logs. Solving the issue creating a separate temp local file
for handle_unreachable and extract timeago in a variable in order to use it to
determine the age of the job. This keeps clear now the out variable which
should only contain the autoinst-log+reason.

https://progress.opensuse.org/issues/166772

Signed-off-by: ybonatakis <ybonatakis@suse.com>
Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>