Avoid submitting packages to Factory that we haven't actually tested yet
The packages on `devel:openQA` might change while openQA-in-openQA tests
are executing. So far we submit packages directly from `devel:openQA`
after openQA-in-openQA have passed. This means we don't neccassarily submit
the version of the packages we have actually tested.

With this change we "release" a snapshot of `devel:openQA` to the staging
repository `devel:openQA:testing` when triggering tests and use that within
openQA-in-openQA tests (see
https://github.com/os-autoinst/os-autoinst-distri-openQA/pull/215). We then
only submit from this staging repository when tests have passed. This way
we only submit what we've tested despite changes on `devel:openQA`.

Packages in the staging repository are cleaned up when openQA-in-openQA
tests fail or after packages have been submitted. This trigger script will
only release a new version into the staging project if it has been cleaned
up ensuring its contents stay the same while tests/submissions are pending.

This commit also adds the new script `cleanup-obs-project`. It conveniently
deletes all packages in an OBS project without deleting the project itself.
We might want to add it on Jenkins as post-failure action for monitoring
and submissions to avoid leftovers. (It is also useful to test the deletion
code on its own.)

Related ticket: https://progress.opensuse.org/issues/167395