Avoid error message for failing command
    /home/runner/work/scripts/scripts/.bpan/.bpan/lib/test-tap.bash: ERROR: line 168

Tests pass, it's just printing that error.

I'm not sure why that's needed here as `try` should take care of it, but
I think it's because the error for calling investigate_issue without
an argument is "Unbound variable" and that might not be catched the same way as
other error messages. I might ask the test-tap author about it.