Allow triggering openQA-in-openQA tests if submission is pending
We can schedule the tests anyway; we only should not override the staging
repo and follow up with a submission. This way we see problems with openQA-
in-openQA tests early - even when submissions are pending for a while.

To indicate that we should not do a submission, this change creates the
file `job_post_skip_submission` in the current working directory which is
picked up by e.g. Jenkins as artifact via the glob `job_post_*` and can be
used by subsequent jobs. It is important to configure that subsequent jobs
use the artifacts from the preceding job that triggered them and *not* the
last successful job.

Related ticket: https://progress.opensuse.org/issues/167395