Throttle the number of SRs to Factory in submit script itself
So far the throttling is done on Jenkins but it has the disadvantage that
submission jobs skipped because there is already a pending submission also
count towards the limit. With this change we simply check whether there are
any recent pending SRs in `os-autoinst-obs-auto-submit` itself to avoid
this problem.

Related ticket: https://progress.opensuse.org/issues/167395