Prevent stack trace from subprocess.check_output
This commit makes sure that the python stack trace is handled appropiately in
main function, and logs the error neverthless, for easy tracking. It uses
SubprocessError is order to get a general variaty of exceptions from
subprocess. Changes include redirection of the stderr from the subprocess.
the stderr and stdout are compined. This makes easy to test and log the
responds.

https://progress.opensuse.org/issues/174601