Add function to submit openQA packages to openSUSE:Backports
- Separate first submission to factory from anything
else. By default use the submit_target variable which targets openSUSE:Factory
- Refactor factory_request to take extra variables and make the query more
abstract and modular
- Encapsulate logic in a main and break it into smaller logical functions

The automatic submission is designed to sumbmit to the Factory using
staging repo. However after that there is no need of this. The packages can be
submitted to any other project from factory once they are accepted in there.

https://progress.opensuse.org/issues/127037