Create simultaneous submission of the packages to obs
- Refactor factory_request (renamed to get_obs_sr_id) to take extra variables
and make the query more abstract and modular. rename it to `get_obs_sr_id` to
represent clear its purpose
- Keep the submit_target as it is used in all the phases before the
update_package. But its logic is changed a bit as there is a new variable
`targets` representing the final submission targets. `targets` are combined
from `submit_target` and `submit_target_extra`
- Encapsulate logic in a main and break it into smaller logical functions
- Capture and report failed packages

The automatic submission is designed to submit to the Factory using
staging repo. However, after that there is no need for this. The packages can be
submitted to any other project from factory once they are accepted in
there. Once the `update_package` is called, it is used to perform the submit
request to factory, now it will loop through a list of `$targets` and create a
request (aka osc sr).

https://progress.opensuse.org/issues/127037