Test the variable expansion on the submit_target_extra
Test cases for the submit_target_extra variable:
- the default project is set if it is matched with the `osc dists` output
- exit if `osc dists` doesnt found any match in the output of the extracted
project of submit_target_extra
- the latest version is used from the `osc dists` output

Notice that the tested function is using and depends on `submit_target_extra_project`
and the output is assigned to the `submit_target_extra`(the default value)). However
only the underline function is tested. Not `submit_target_extra` directly

https://progress.opensuse.org/issues/127037