Test the variable expansion on the submit_target_extra
This is an ungly test which wrap the sequence of lines which used to expand
the submit_target_extra in order to find the latest. These checks ensure that
the most common values behaves as expected.

It is poor in sense that it doesnt call the code but rather copy part of it
and encapsulates it in a function.