Fix regression in the os-autoinst-obs-auto-submit
This is a regression from
https://github.com/os-autoinst/scripts/commit/360f247977c724e7b4aed8fd71b0fcb11e48e2e4
where the `$staging_project` does *not* overide `$src_project` before the
`handle_auto_submit` but instead, it makes it take place after the
function. That had side effect in the behavior of the script.
To make the critical conditions intuitive, When the `src_project` is overidden by the
`${src_project}:testing` the `must_cleanup_staging` is activated to enable the
cleanup of staging.