Fix expansion for submit_target_extra
First the expansion is missing $ at the beginning and it makes it assign the
expression as a string. Additional ensure that the
get_submit_target_extra_latest_version is called as code and returns its
value.

Also the test is updated to source the script later after the osc mocking
because it is likely it tries to execute the function too early with the osc
which tries to find but it does not exist in the environment.

https://progress.opensuse.org/issues/127037