Pick latest version only when Update path exists
Use the `osc meta` in combination with `osc dists` to submit to the most
recent :Update repo. Likely no need for more than two distribution to find the
desired one. `osc meta` returns 404 if the prj does not exist and this should
give a clue to pick the latest Update project even if it is included in `osc
dists`

https://progress.opensuse.org/issues/127037