Use fixed Backports project for submission
The latest Leap, openQA is using is 15.6 which is the _base_ for
SLE-15-SP6. SLE-15-SP7 lacks packages it is not supported for now. As such
the *get_submit_target_extra_latest_version* is not useful here as it returns
the latest version with :Update project which at the moment that would be
SLE-15-SP7.

https://progress.opensuse.org/issues/127037