Remove unused get_submit_target_extra_latest_version
With the `Use fixed Backports project for submission` in place the function
which implements automatic discovery of the latest suitable version is not
needed at the moment.

Revert this to restore the function and take a look in the `Use fixed
Backports project for submission` commit to enable the functionality again.