Avoid triggering openqa_in_openqa job with old builds
`trigger_openqa_in_openqa` is run multiple times per day, selecting the most
recent published Tumbleweed build. Tumbleweed is published more or less in a
frequency of 1 to 2 days. maybe it doesnt make sense to run jobs due to the
absence of the assets which can happen as part of the cleanup to the old
images.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>