Allow users to turn off investigation runs
Users might need to turn off auto investigation runs for various of
reasons. As of now it was not easy to turn off the investigations runs
quickly or the current solution lacks the granularity.
In this case we need to turn off the runs for a specific flavor and in a
specific job group only. This parameter can be placed in an openQA
medium definition.